news/scripts/requirements.txt
args==0.1.0
astroid==2.3.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
clint==0.5.1
coverage==5.0.4
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.9
isort==4.3.21
jieba3k==0.35.1
lazy-object-proxy==1.4.3
lxml==4.5.0
mamba==0.10
mccabe==0.6.1
mock==4.0.2
mockito==1.2.1
newspaper3k==0.2.8
nltk==3.4.5
Pillow==7.0.0
pylint==2.4.4
pymongo==3.10.1
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
requests-file==1.4.3
six==1.14.0
soupsieve==2.0
sure==1.4.11
tinysegmenter==0.3
tldextract==2.2.2
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
